20220204 button gpio modfy

+++ stm32f769i_disco_fw/src/common/hw/include/gpio.h
@@ -24,7 +24,7 @@
 
 
 bool gpioInit(void);
-bool gpioPinMode(uint8_t ch, uint8_t mode);
+bool gpioPinMode(uint8_t ch, uint32_t mode);
 void gpioPinWrite(uint8_t ch, bool value);
 bool gpioPinRead(uint8_t ch);
 void gpioPinToggle(uint8_t ch);
